pyproject.toml
[project]
name = "ecommerce-pro"
version = "0.1.0"
description = "E-commerce backend built with Django 5.1+, DRF, and JWT authentication, designed for React frontend integration"
requires-python = ">=3.11"
dependencies = [
    "django-cors-headers>=4.7.0",
    "django>=5.1.7",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
]
